requirements.txt
appdirs==1.4.4
appnope==0.1.2
backcall==0.2.0
black==21.5b2
click==8.0.1
dataclasses==0.8
decorator==5.0.9
Django==2.2
djangorestframework==3.12.4
importlib-metadata==4.5.0
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
mypy-extensions==0.4.3
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
zipp==3.4.1
